requirements.txt
django==5.2
django-admin-interface==0.30.0
django-jazzmin==3.0.1

langchain
langchain-openai

django-crispy-forms==2.4
crispy-bootstrap5==2025.4

pillow==11.2.1
python-dotenv==1.1.0

gunicorn==23.0.0

celery==5.5.3

flake8==7.3.0
black==25.1.0
# isort==6.0.1
autoflake==2.3.1
